Treat whitespace strings as empty() in ValueChecks.
* Treats strings that are only whitespace as "empty" in the funtion `ValueChecks.empty()`.

Note that this change could be quite controversial. However, consider the naming of `empty` and what we actually mean by it. Let's say this is used to validate whether a user has entered anything into a textbox or not. The user would probably agree that she did not enter anything if it only contains a single space. Another example would be to validate whether some http response from a service actually contained anything. If there's only whitespace returned from the service, it's more likely we would think of that response as "empty" than "contains an actual response".