Cart: Discount per item
Introduces the possiblility to have a discount attatched to an item.
This discount will only be applied on the local itemline: the resulting
amount will be considered as it's gross amount, which is the value used
when doing all cart-wide calculations such as global discount
distribution etc.

TODO: Conclude if we want to add a method on the Cart for summarizing
*all* discounts, including the oned local to item lines.