UUIDFactory: Add method for parsing both string versions
As we're moving away from using the base64 representation of a `UUID`,
we still want to allow the old format in API's and as input parameters.
Testing for the two different versions everywhere makes the code ugly
and repetitive.

This change simply adds a parse method that will accept both versions of
the serialization format.

Also, the access modifier of the method `fromBase64String` is changed to
default, as exposing this method is now only redundant.