Local Queue for failed SNS publishing
Implementation of a local Queue based on H2 DB. This allows
us to choose between an in-memory queue or a file based queue.
The H2 db is quite performant and should suffice for us to begin
with. Using the H2 datbase makes the FIFO queue implemenation easy.
An improvement to speed things up is to implement a simple FIFO Q
as a RandomAccessFile, but this is a improvement that I don't
think we need at the moment

I wanted the Local Queue to be injected into SUN code without too
much change needed, hence I have baked the Queue into the
MessagePublisher interface such that it can be used in SUN:s
MessagingManager class without too much change. This however made
the task a bit bigger, in fact most changes in this commit is note
connected to the Q itself but the surrounding classes.

If we fail to send a message via a MessagePublisher it can be due
to many reasons but the two big ones are Network issues towords SNS
and serialization issues for the messages being sent. If there is a
network problem then we know for sure that once the network is up we
will be able to send the message. If there is a serialization issue we
need to make a code release in order to fix that. In any case messages
that cannot be handled are removed from the main Q after X number of
retries and pushed onto a DLQ queue. The DLQ is just a normal Q that
is polled less frequently. This feature will make it possible to remove
events from the top of the Q that cannot be handled and risk blocking
events behind them from being processed. There is however a problem
with the not serializable case as these messages might contain something
that is wrong and thus stay in the DLQ forever and block the DLQ
eventually. A time based eviction strategy is probably needed for
the DLQ in order to not fill it up completely.

The introduction of the Q also introduces the fact that we need
threads to poll both the Q and DLQ in order to push the localy stored
events onto the SNS:s again. As there is one Q per SNS topic we will get
quite some queues to pull, to not need one thread per topic a
ExecutorServices is used in order to control the number of threads
polling Q:s

The class TaskQueueMessagePublisherFactory is the factory that
creates a MessagePublisher backed with a Q and destination
MessagePublisher.

Outstanding from this commit is the time based eviction from the
Q in order to not block the top of the Q forever in case of many
bad events being sent into the Q.