izettle-java: InstantTruncator - truncating and adding time to `Instant`
Unfortunately, the new java.time API doesn't allow for truncating units
larger than days, which is a show stopper for quite a few business
problems.

This utility class simply mimics the truncation behaviour of the
`CalendarTruncator` in the toolbox: just wrapping around an `Instant`
instead of a `Date`. Also, the `plus` and `minus` methods of `Instant`
also doesn't allow for units larger than a 24H, so those are added as
well for utility (so that the caller doesn't have to convert to and
from a `ZonedDateTime`).