izettle-java: Interval - Adds new Interval class
Unfortunately, the java8 time API doesn't have a proper interval class.
The only such classes in standard java is `Duration` and `Period`, but
they only hold information about the length of time: not the instant on
the timeline.
So, a proper interval class should really have been part of the standard
`java.time` API.

This class does that. It's a simple class consisting of two `Instant`s.
It's start inclusive, and end exclusive. Also, it's got some utility
methods, such as `overlap`, `abuts` and `contains`.