Fix how line item discounts are calculated (#177)
Line item discounts are now calculated on lines rounded gross value
instead of it's non-rounded value. This is to avoid situations where the
discount value is calculated on an exact value, and then rounded in a
direction that differs from how the final line items value are rounded:
effectively disproportionately changing the percentage of the discount
in relation to the gross amount.

Added tests that fail before this code change.