TLVDecoder remove array alloc during recursion.
This PR removes the new byte array allocation during recursive
calls. It reuses the ByteBuffer when the value of a tag is a new
TLV byte array. It keeps the array allocation when the value of a
tag is _just_ a "value", this should keep array allocations at a
minumum and my very un-educated guess is that the big arrays are the
ones where the value of a tag is a TLV.